vtryonzip/vtryon/requirements.txt
Django==3.2
requests==2.24.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
django-cors-headers
pandas
nltk
regex
scikit-learn
pymongo
cython
cmfrec
apyori
psycopg2-binary>=2.8

joblib